requirements.txt
blessed==1.20.0
boto3==1.36.16
botocore==1.36.16
chalice==1.31.4
click==8.1.8
editor==1.6.6
inquirer==3.4.0
jmespath==1.0.1
pillow==11.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
readchar==4.2.1
runs==1.2.2
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
xmod==1.8.1
